refactor(meeting_repl/repl): split tests to sibling tests_repl.rs (#1266)
- src/meeting_repl/repl.rs: 548 → 304 LOC
- New src/meeting_repl/tests_repl.rs (241 LOC)
- mod registration

Pure refactor: 9 tests pass.
